fix: get rid of  `-Wsign-compare` warnings

+++ src/messages/layouts/MessageLayoutElement.cpp
@@ -514,7 +514,7 @@ int TextLayoutElement::getXFromIndex(size_t index)
     else if (index < static_cast<size_t>(this->getText().size()))
     {
         int x = 0;
-        for (int i = 0; i < index; i++)
+        for (size_t i = 0; i < index; i++)
         {
             x += metrics.horizontalAdvance(this->getText()[i]);
         }
